class4s: fix the compiler error about unknown greeting()
This fixes the first problem, but now there's the next one -- that's the same
that happened when using C: The main() function tries to call greeting(),
but there's no actual greeting() function to be called/linked, since that's
in a different source file we're not even compiling.

```
kais-mbp:cxx kai$ ./build.sh
Undefined symbols for architecture x86_64:
  "__Z8greetingv", referenced from:
      _main in ccV2Z1sO.o
ld: symbol(s) not found for architecture x86_64
collect2: error: ld returned 1 exit status
```

A slight difference is that the symbol of the missing function `__Z8greetingv`
looks kinda odd. It's a "mangled" C++ function name.