class4s: personalize the greeting
Just like in the C case, we personalize the greeting, but we only change the
greeting() function itself to take a name, but forget to make that change
where we call the function.

As opposed to C, which ended up crashing, in C++ we get a link error, since
C++ does distinguish between `greeting(void)` and `greeting(const char*)`,
and we're calling the former but it does not exist, so the linker complains
about it. (Which is better than a random crash).

```
kais-mbp:cxx kai$ ./build.sh
Undefined symbols for architecture x86_64:
  "__Z8greetingv", referenced from:
      _main in hello.o
ld: symbol(s) not found for architecture x86_64
collect2: error: ld returned 1 exit status
```