class8s: automate the vector_dot() test
I'm just using the basic "assert()" way here, since this is C, so I cannot
(easily) use googletest.

One drawback to be aware of is that when compiling code in Release mode in
cmake, it adds the `-DNDEBUG` flag, which turns off assert()s entirely, so
in Release mode, this kind of test doesn't really work.